Removed assignments and started to re-write handler

+++ baker_database/database.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# For the room, assignment and robot data information
+# For the room and robot data information
 import database_classes
 # For time calculations
 from datetime import datetime, date, time, timedelta
@@ -20,8 +20,6 @@ from ipa_building_msgs.msg import *
 
 # Database class
 class Database():
-	# Assignments
-	assignments_ = []
 	# Rooms
 	rooms_ = []
 	# Global map data
@@ -32,9 +30,7 @@ class Database():
 	global_settings_ = None
 	# File names
 	rooms_filename_ = ""
-	assignments_filename_ = ""
 	tmp_rooms_filename_ = ""
-	tmp_assignments_filename_ = ""
 	global_settings_filename_ = ""
 	robot_properties_filename_ = ""
 	global_map_data_filename_ = ""
@@ -94,38 +90,35 @@ class Database():
 		test_room.room_issues_.append(issue1)
 		test_room.room_issues_.append(issue2)
 		test_room.room_id_ = 42
-		test_room.last_cleanup_successful_ = False
-		test_room.last_successful_cleaning_date_ = None
+		#test_room.last_cleanup_successful_ = False
+		#test_room.last_successful_cleaning_date_ = None
 		test_room.room_map_ = None
 		test_room.room_center_coords_ = None
 		test_room.room_name_ = "Kitchen"
 		# Add room to the room list
 		self.rooms_.append(test_room)
-	
-	
-	# Create a test assignment dictionary
-	def createTestAssignmentDict(self):
-		assignment_dict = {}
-		assignment_dict[42] = {
-			"assignment_name": "Assignment Number Two",
-			"scheduled_rooms_cleaning": [5,6,7,8],
-			"last_completed_clean": datetime.now().strftime("%Y-%m-%d_%H:%M")
-		}
-		return assignment_dict
-
-
-	# Add a test assignment object to the assignments_ list
-	def createTestAssignmentObject(self):
-		test_assignment = database_classes.AssignmentItem()
-		test_assignment.assignment_name_ = "Assignment Number One"
-		test_assignment.scheduled_rooms_cleaning_ = [1,2,3,4]
-		test_assignment.last_completed_clean_ = datetime.now()
-		self.assignments_.append(test_assignment)
 
 # =========================================================================================
 # Private methods
 # =========================================================================================
 
+	@staticmethod
+	def datetimeToString(datetime_date):
+		return datetime_date.strftime("%Y-%m-%d_%H:%M")
+
+	@staticmethod
+	def stringToDatetime(string_date):
+		return datetime.strptime(string_date, "%Y-%m-%d_%H:%M")
+
+	@staticmethod
+	def point32ToArray(point32_point):
+		return [point32_point.x, point32_point.y, point32_point.z]
+
+	@staticmethod
+	def arrayToPoint32(array_point):
+		return Point32(x=array_point[0], y=array_point[1], z=array_point[2])
+	
+	
 	def updateGlobalSettings(self, dict):
 		self.global_settings_ = database_classes.GlobalSettings()
 		self.global_settings_.shall_auto_complete_ = dict.get("shall_auto_complete")
@@ -247,18 +240,32 @@ class Database():
 			current_room.room_surface_area_ = dict.get(room_key).get("room_surface_area")
 			# Get the room trashcan count
 			current_room.room_trashcan_count_ = dict.get(room_key).get("room_trashcan_count")
-			# Get the last successful cleaning date if there is any, otherwise set None
-			date_str = dict.get(room_key).get("last_successful_cleaning_date")
-			if (date_str != None):
-				current_room.last_successful_cleaning_date_ = datetime.strptime(date_str, "%Y-%m-%d_%H:%M")
-			else:
-				current_room.last_successful_cleaning_date_ = None
-			# Get the last successful trashcan date if there is any, otherwise set None
-			date_str = dict.get(room_key).get("last_successful_trashcan_date")
-			if (date_str != None):
-				current_room.last_successful_trashcan_date_ = datetime.strptime(date_str, "%Y-%m-%d_%H:%M")
-			else:
-				current_room.last_successful_trashcan_date_ = None
+			# Get the days where the room has to be cleaned in a specified way
+			current_room.room_scheduled_days_ = dict.get(room_key).get("room_scheduled_days")
+			
+			## Get the last successful cleaning date if there is any, otherwise set None
+			#date_str = dict.get(room_key).get("last_successful_cleaning_date")
+			#if (date_str != None):
+			#	current_room.last_successful_cleaning_date_ = datetime.strptime(date_str, "%Y-%m-%d_%H:%M")
+			#else:
+			#	current_room.last_successful_cleaning_date_ = None
+			## Get the last successful trashcan date if there is any, otherwise set None
+			#date_str = dict.get(room_key).get("last_successful_trashcan_date")
+			#if (date_str != None):
+			#	current_room.last_successful_trashcan_date_ = datetime.strptime(date_str, "%Y-%m-%d_%H:%M")
+			#else:
+			#	current_room.last_successful_trashcan_date_ = None
+			
+			# Get the list with the datestamps
+			string_datestamp_list = dict.get(room_key).get("room_cleaning_datestamps")
+			datestamps = []
+			for datestamp in string_datestamp_list:
+				if (datestamp != None):
+					datestamps.append(self.stringToDatetime(datestamp))
+				else:
+					datestamps.append(None)
+			current_room.room_cleaning_datestamps_ = datestamps
+			
 			# Append current room object to the rooms_ list
 			self.rooms_.append(current_room)
 
@@ -291,42 +298,38 @@ class Database():
 						}
 					else:
 						print "[FATAL]: An element in issues array is not an issue object!"
-				# Fill in the last successful cleaning date if there is any, otherwise fill in None
-				if (current_room.last_successful_cleaning_date_ != None):
-					date_str_cleaning = current_room.last_successful_cleaning_date_.strftime("%Y-%m-%d_%H:%M")
-				else:
-					date_str_cleaning = None
-				# Fill in the last successful trashcan date if there is any, otherwise fill in None
-				if (current_room.last_successful_trashcan_date_ != None):
-					date_str_trashcan = current_room.last_successful_trashcan_date_.strftime("%Y-%m-%d_%H:%M")
-				else:
-					date_str_trashcan = None
+				
+				## Fill in the last successful cleaning date if there is any, otherwise fill in None
+				#if (current_room.last_successful_cleaning_date_ != None):
+				#	date_str_cleaning = current_room.last_successful_cleaning_date_.strftime("%Y-%m-%d_%H:%M")
+				#else:
+				#	date_str_cleaning = None
+				## Fill in the last successful trashcan date if there is any, otherwise fill in None
+				#if (current_room.last_successful_trashcan_date_ != None):
+				#	date_str_trashcan = current_room.last_successful_trashcan_date_.strftime("%Y-%m-%d_%H:%M")
+				#else:
+				#	date_str_trashcan = None
+				
+				# Fill in the datestamps
+				datestamp_list = []
+				for datestamp in current_room.room_cleaning_datestamps_:
+					if datestamp != None:
+						datestamp_list.append(self.datetimeToString(datestamp))
+					else:
+						datestamp_list.append(None)
 				# Fill in the room information
 				if ((current_room.room_information_in_meter_ != None) and (current_room.room_information_in_pixel_ != None)):
-					px_center_x = current_room.room_information_in_pixel_.room_center.x
-					px_center_y = current_room.room_information_in_pixel_.room_center.y
-					px_center_z = current_room.room_information_in_pixel_.room_center.z
-					px_min_x = current_room.room_information_in_pixel_.room_min_max.points[0].x
-					px_min_y = current_room.room_information_in_pixel_.room_min_max.points[0].y
-					px_min_z = current_room.room_information_in_pixel_.room_min_max.points[0].z
-					px_max_x = current_room.room_information_in_pixel_.room_min_max.points[1].x
-					px_max_y = current_room.room_information_in_pixel_.room_min_max.points[1].y
-					px_max_z = current_room.room_information_in_pixel_.room_min_max.points[1].z
-					meter_center_x = current_room.room_information_in_meter_.room_center.x
-					meter_center_y = current_room.room_information_in_meter_.room_center.y
-					meter_center_z = current_room.room_information_in_meter_.room_center.z
-					meter_min_x = current_room.room_information_in_meter_.room_min_max.points[0].x
-					meter_min_y = current_room.room_information_in_meter_.room_min_max.points[0].y
-					meter_min_z = current_room.room_information_in_meter_.room_min_max.points[0].z
-					meter_max_x = current_room.room_information_in_meter_.room_min_max.points[1].x
-					meter_max_y = current_room.room_information_in_meter_.room_min_max.points[1].y
-					meter_max_z = current_room.room_information_in_meter_.room_min_max.points[1].z
-					room_information_in_pixel_list = [[px_center_x, px_center_y, px_center_z], [px_min_x, px_min_y, px_min_z], [px_max_x, px_max_y, px_max_z]]
-					room_information_in_meter_list = [[meter_center_x, meter_center_y, meter_center_z], [meter_min_x, meter_min_y, meter_min_z], [meter_max_x, meter_max_y, meter_max_z]]
+					px_center_list = self.point32ToArray(current_room.room_information_in_pixel_.room_center)
+					px_min_list = self.point32ToArray(current_room.room_information_in_pixel_.room_min_max.points[0])
+					px_max_list = self.point32ToArray(current_room.room_information_in_pixel_.room_min_max.points[1])
+					room_information_in_pixel_list = [px_center_list, px_min_list, px_max_list]
+					meter_center_list = self.point32ToArray(current_room.room_information_in_meter_.room_center)
+					meter_min_list = self.point32ToArray(current_room.room_information_in_meter_.room_min_max.points[0])
+					meter_max_list = self.point32ToArray(current_room.room_information_in_meter_.room_min_max.points[1])
+					room_information_in_meter_list = [meter_center_list, meter_min_list, meter_max_list]
 				else:
 					room_information_in_pixel_list = [None, None, None]
 					room_information_in_meter_list = [None, None, None]
-
 				# Fill the dictionary with the data
 				room_dict[str(current_room.room_id_)] = {
 					"room_id": current_room.room_id_,
@@ -335,8 +338,8 @@ class Database():
 					"room_floor_id": current_room.room_floor_id_,
 					"room_building_id": current_room.room_building_id_,
 					"room_territory_id": current_room.room_territory_id_,
-					"last_successful_cleaning_date": date_str_cleaning,
-					"last_successful_trashcan_date": date_str_trashcan,
+					#"last_successful_cleaning_date": date_str_cleaning,
+					#"last_successful_trashcan_date": date_str_trashcan,
 					"room_issues": issues_dict,
 					"room_map": current_room.room_map_,
 					"room_information_in_pixel": room_information_in_pixel_list,
@@ -344,70 +347,13 @@ class Database():
 					"room_surface_type": current_room.room_surface_type_,
 					"room_cleaning_method": current_room.room_cleaning_method_,
 					"room_surface_area": current_room.room_surface_area_,
-					"room_trashcan_count": current_room.room_trashcan_count_
+					"room_trashcan_count": current_room.room_trashcan_count_,
+					"room_scheduled_days": current_room.room_scheduled_days_,
+					"room_cleaning_datestamps": datestamp_list
 				}
 			else:
 				print "[FATAL]: An element in rooms_ array is not a room object!"
 		return room_dict
-
-
-	# Make assignments_ contain all the rooms stated in the dict parameter
-	def updateAssignmentsList(self, dict):
-		self.assignments_ = []
-		for assignment_key in dict:
-			current_assignment = database_classes.AssignmentItem()
-			# Get the name of the previous assignment
-			current_assignment.prev_assignment_ = dict.get(assignment_key).get("prev_assignment")
-			# Get the name of the assignment
-			current_assignment.assignment_name_ = dict.get(assignment_key).get("assignment_name")
-			# Get the week type the assignment is active
-			current_assignment.assignment_week_type_ = dict.get(assignment_key).get("assignment_week_type")
-			# Get week day the assignment is active
-			current_assignment.assignment_week_day_ = dict.get(assignment_key).get("assignment_week_day")
-			# Get the list of room IDs for cleaning which are scheduled in this assignment
-			current_assignment.scheduled_rooms_cleaning_ = dict.get(assignment_key).get("scheduled_rooms_cleaning")
-			# Get the list of room IDs for trashcan which are scheduled in this assignment
-			current_assignment.scheduled_rooms_trashcan_ = dict.get(assignment_key).get("scheduled_rooms_trashcan")
-			# Get the RoomItem representation behind the ID for all rooms to clean
-			for room_id in current_assignment.scheduled_rooms_cleaning_:
-				current_assignment.scheduled_rooms_cleaning_data_.append(self.getRoom(room_id))
-			# Get the RoomItem representation behind the ID for all rooms to empty the trashcan
-			for room_id in current_assignment.scheduled_rooms_trashcan_:
-				current_assignment.scheduled_rooms_trashcan_data_.append(self.getRoom(room_id))
-			# Get a date string or None if there is no date
-			date_str = dict.get(assignment_key).get("last_completed_clean")
-			if (date_str != None):
-				current_assignment.last_completed_clean_ = datetime.strptime(date_str, "%Y-%m-%d_%H:%M")
-			else:
-				current_assignment.last_completed_clean_ = None
-			# Append current_assignment to assignments_ list
-			self.assignments_.append(current_assignment)
-
-
-	# Get a dictionary representation of assignments_
-	def getAssignmentsDictFromAssignmentsList(self):
-		assignment_dict = {}
-		for current_assignment in self.assignments_:
-			# Check if current_assignment is an AssignmentItem object
-			if (isinstance(current_assignment, database_classes.AssignmentItem)):
-				# Get the last completed clean date if there is any, otherwise set it as None
-				if (current_assignment.last_completed_clean_ != None):
-					date_str = current_assignment.last_completed_clean_.strftime("%Y-%m-%d_%H:%M")
-				else:
-					date_str = None
-				# Fill the dictionary with the data
-				assignment_dict[current_assignment.assignment_name_] = {
-					"prev_assignment": current_assignment.prev_assignment_,
-					"assignment_name": current_assignment.assignment_name_,
-					"assignment_week_type": current_assignment.assignment_week_type_,
-					"assignment_week_day": current_assignment.assignment_week_day_,
-					"scheduled_rooms_cleaning": current_assignment.scheduled_rooms_cleaning_,
-					"scheduled_rooms_trashcan": current_assignment.scheduled_rooms_trashcan_,
-					"last_completed_clean": date_str
-				}
-			else:
-				print "[FATAL]: An element in assignments_ array is not an assignment object!"
-		return assignment_dict
 
 
 
@@ -418,9 +364,7 @@ class Database():
 	# Constructor method
 	def __init__(self, extracted_file_path=""):
 		self.extracted_file_path = extracted_file_path
-		self.assignments_filename_ = self.extracted_file_path + str("resources/json/assignments.json")
 		self.rooms_filename_ = self.extracted_file_path + str("resources/json/rooms.json")
-		self.tmp_assignments_filename_ = self.extracted_file_path + str("resources/json/tmp_assignments.json")
 		self.tmp_rooms_filename_ = self.extracted_file_path + str("resources/json/tmp_rooms.json")
 		self.robot_properties_filename_ = self.extracted_file_path + str("resources/json/robot_properties.json")
 		self.global_settings_filename_ = self.extracted_file_path + str("resources/json/global_settings.json")
@@ -431,8 +375,6 @@ class Database():
 	def discardTemporalDatabase(self):
 		if (os.path.isfile(self.tmp_rooms_filename_) == True):
 			os.remove(str(self.tmp_rooms_filename_))
-		if (os.path.isfile(self.tmp_assignments_filename_) == True):
-			os.remove(str(self.tmp_assignments_filename_))
 		self.loadDatabase()
 
 	# Load database data from files
@@ -444,13 +386,6 @@ class Database():
 			file = open(self.rooms_filename_, "r").read()
 		rooms_dict = json.loads(file)
 		self.updateRoomsList(rooms_dict)
-		# Load the assignment data
-		if (os.path.isfile(self.tmp_assignments_filename_) == True):
-			file = open(self.tmp_assignments_filename_, "r").read()
-		else:
-			file = open(self.assignments_filename_, "r").read()
-		assignments_dict = json.loads(file)
-		self.updateAssignmentsList(assignments_dict)
 		# Load the robot properties
 		file = open(self.robot_properties_filename_, "r").read()
 		robot_properties_dict = json.loads(file)
@@ -464,8 +399,10 @@ class Database():
 		global_map_data_dict = json.loads(file)
 		self.updateGlobalMapData(global_map_data_dict)
 
-	# Save database data to files
-	def saveDatabase(self, temporal=True):
+
+
+	# Save room database data to files
+	def saveRoomDatabase(self, temporal=True):
 		# Save the room data
 		rooms_dict = self.getRoomsDictFromRoomsList()
 		rooms_text = json.dumps(rooms_dict, indent=4, sort_keys=True)
@@ -474,31 +411,9 @@ class Database():
 		else:
 			file = open(self.rooms_filename_, "w")
 		file.write(rooms_text)
-		# Save the assignment data
-		assignments_dict = self.getAssignmentsDictFromAssignmentsList()
-		assignments_text = json.dumps(assignments_dict, indent=4, sort_keys=True)
-		if (temporal == True):
-			file = open(self.tmp_assignments_filename_, "w")
-		else:
-			file = open(self.assignments_filename_, "w")
-		file.write(assignments_text)
-		"""
-		# Save global settings
-		global_settings_dict = self.getGlobalSettingsDictFromGlobalSettings()
-		global_settings_text = json.dumps(global_settings_dict, indent=4, sort_keys=True)
-		file = open(self.global_settings_filename_, "w")
-		file.write(global_settings_text)
-		"""
-		# Save robot properties? No.....
-		# [...]
-		# Save global map data? No.....
-		# [...]
-		# Remove all temporal files, if wanted
-		if (temporal == False):
-			if (os.path.isfile(self.tmp_rooms_filename_) == True):
-				os.remove(str(self.tmp_rooms_filename_))
-			if (os.path.isfile(self.tmp_assignments_filename_) == True):
-				os.remove(str(self.tmp_assignments_filename_))
+		if ((temporal == False) and (os.path.isfile(self.tmp_rooms_filename_) == True)):
+			os.remove(str(self.tmp_rooms_filename_))
+
 
 
 	# Retreive a room by providing a room_id
@@ -509,32 +424,6 @@ class Database():
 				result = self.rooms_[i]
 		return result
 
-	# Retreive a room by providing a position id and a floor id
-	def getRoomByPosFloor(self, pos_id, floor_id):
-		result = None
-		for room in self.rooms_:
-			if ((room.room_position_id_ == pos_id) and (room.room_floor_id_ == floor_id)):
-				result = room
-				break
-		return result
-
-	# Retreive an assignment by providing a week day and week type
-	def getAssignmentByWeekTypeDay(self, week_type, week_day):
-		result = None
-		for assignment in self.assignments_:
-			if ((assignment.assignment_week_type_ == week_type) and (assignment.assignment_week_day_ == week_day)):
-				result = assignment
-				break
-		return result
-
-	# Retreive an assignment by providing the name of it
-	def getAssignmentByName(self, name):
-		result = None
-		for assignment in self.assignments_:
-			if (assignment.assignment_name_ == name):
-				result = assignment
-				break
-		return assignment
 
 
 """
@@ -547,7 +436,6 @@ class Database():
 db = Database()
 db.loadDatabase()
 #db.createTestRoomObject()
-#db.createTestAssignmentObject()
 
 # Play around with the containing data
 print db.getRoom(21).room_issues_[0].issue_id_
